requirements.txt
pandas
datetime
pillow
psd_tools
ttkthemes
parsel
xlsxwriter
xlrd
openpyxl
shutil
os

